Bircok Degişiklik yapıldı.

+++ post/views.py
@@ -1,8 +1,8 @@
-from django.shortcuts import render, HttpResponse, \
-    get_object_or_404, HttpResponseRedirect,redirect
+from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect, Http404
 from .models import Post
 from .form import PostForm
-
+from django.contrib import messages
+from django.utils.text import slugify
 
 # Create your views here.
 
@@ -12,8 +12,8 @@ def post_index(request):
     return render(request, 'post/index.html', {'posts': posts})
 
 
-def post_detail(request, id):
-    posts = get_object_or_404(Post, id=id)
+def post_detail(request, slug):
+    posts = get_object_or_404(Post, slug=slug)
     contex = {
         'posts': posts,
     }
@@ -21,6 +21,17 @@ def post_detail(request, id):
 
 
 def post_create(request):
+    if not request.user.is_authenticated:
+        return Http404()
+    forms = PostForm(request.POST or None, request.FILES or None)
+    if forms.is_valid():
+        post = forms.save()
+
+        messages.success(request, 'Başarılı Bir Şekilde Oluşturdunuz.', extra_tags='mesaj-basarili')
+        return HttpResponseRedirect(post.get_absolute_url())
+    context = {'forms': forms}
+    return render(request, 'post/form.html', context)
+
     # if request.method == "POST":
     #     # Formdan Gelen Bilgileri Kaydet
     #     forms = PostForm(request.POST)
@@ -32,26 +43,25 @@ def post_create(request):
 
     # Üstteki Satırlar da Aynı işi yapar
     #  forms = PostForm(request.POST or None) Post dolu geldiyse parametre olarak al demektir.
-    forms = PostForm(request.POST or None)
-    if forms.is_valid():
-        post = forms.save()
-        return HttpResponseRedirect(post.get_absolute_url())
-    context = {'forms': forms}
-    return render(request, 'post/form.html', context)
 
 
-def post_update(request, id):
-    post = get_object_or_404(Post, id=id)
-    forms = PostForm(request.POST or None, instance=post)
+def post_update(request, slug):
+    if not request.user.is_authenticated:
+        return Http404()
+    post = get_object_or_404(Post, slug=slug)
+    forms = PostForm(request.POST or None, request.FILES or None, instance=post)
     if forms.is_valid():
         forms.save()
+        messages.success(request, 'Başarılı Bir Şekilde Güncellediniz.')
         return HttpResponseRedirect(post.get_absolute_url())
     context = {'forms': forms}
     return render(request, 'post/form.html', context)
     # return HttpResponse('Burası Post update')
 
 
-def post_delete(request, id):
-    post = get_object_or_404(Post, id=id)
+def post_delete(request, slug):
+    if not request.user.is_authenticated:
+        return Http404()
+    post = get_object_or_404(Post, slug=slug)
     post.delete()
     return redirect('post:index')
